refactor(task-manager): cache key builder component and related tests
